Initial Merges + Savings Matrix + And Added Constraint Functionality
I have merged the large picks ups and deliveries. The most important part of this commit is that I know found (in hindsight a simple method) to sum over the arcs. The savings matrix is reused from the last implementation, which is fine but not optimized, however, imo it is more than sufficient and the various k values are a naive but oke heuristic for this use-case.